fix template compile error
